add fix for iphone mdm off error message to host details
